all code in english

+++ academia/src/main/java/es/ssdd/academia/controllers/ForumController.java
@@ -1,8 +1,6 @@
 package es.ssdd.academia.controllers;
 
-import es.ssdd.academia.entities.Comment;
 import es.ssdd.academia.entities.Course;
-import es.ssdd.academia.services.CommentService;
 import es.ssdd.academia.services.CourseService;
 import es.ssdd.academia.services.ForumService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,7 +20,7 @@ public class ForumController {
     CourseService courseService;
 
     @GetMapping("/{id}/deleteComments/")    //Delete all comments
-    public String deleteComments(@PathVariable int id) {
+    public String deleteComments(@PathVariable long id) {
         Course c = courseService.getOne(id);
         forumService.deleteAllComments(c.getFk_forum());
         return "redirect:/courses/{id}/";
